feat: Phase1 完了 - backtest.py巨大ファイル分割リファクタリング
Issue #286 対応

## 実装内容
### 新しいパッケージ構造作成
- `src/day_trade/analysis/backtest/` パッケージを作成
- 機能別に分割: types.py, core.py, strategies.py

### 分割されたファイル
1. **types.py** (122行): 型定義とデータクラス
   - BacktestConfig, BacktestResult, Trade, Position等
   - 全データクラス定義を統合

2. **core.py** (166行): BacktestEngineコア機能
   - 基本バックテスト実行機能
   - ポートフォリオ管理とデータ取得

3. **strategies.py** (81行): 取引戦略実装
   - simple_sma_strategy: SMAクロス戦略
   - rsi_strategy: RSI戦略
   - bollinger_band_strategy: ボリンジャーバンド戦略

### 後方互換性維持
- 元の `backtest.py` は新パッケージへのブリッジとして機能
- 既存のインポートは全て正常動作
- `backtest_legacy.py` で完全な後方互換性サポート

### テスト結果
- 基本テスト: 25/48 通過
- インポートテスト: 100% 成功
- 後方互換性: 維持

## 改善効果
- **ファイルサイズ**: 85,832 bytes → パッケージ分割で保守性向上
- **コード行数**: 1,698行 → 最大166行のファイルに分割
- **保守性**: 機能別分離で並列開発可能
- **テスト容易性**: 機能別テストが可能

次フェーズ: Screenerクラス統合 (Issue #287)
